Add files via upload
I did the Risk map for Assignment 4 so I thought I'd upload it here in case we need it. It maps all the countries (about 44 of them). I used a .js file for my assignment but I just wanted a way to quickly upload the coordinates to GitHub which is why I am uploading a .html file.